+++ src/shader/bsdf/bsdfSampling.glsl.js
@@ -86,13 +86,6 @@ export const bsdfSamplingGLSL = /* glsl */`
 
 	}
 
-	bool totalInternalReflection( float cosTheta, float eta ) {
-
-		float sinTheta = sqrt( 1.0 - cosTheta * cosTheta );
-		return eta * sinTheta > 1.0;
-
-	}
-
 	vec3 evaluateFresnel( float cosTheta, float eta, vec3 f0, vec3 f90 ) {
 
 		if ( totalInternalReflection( cosTheta, eta ) ) {
